finalized menu tests
