+++ unity/Assets/Tests/PlayModeTests/PcMenuTests/MainMenuLanguageTests.cs
@@ -1,4 +1,4 @@
-using System.Collections;
+﻿using System.Collections;
 using System.Linq;
 using GEAR.Localization;
 using NUnit.Framework;
@@ -9,14 +9,17 @@ using UnityEngine.UI;
 using static Tests.Utilities.Constants;
 using static Tests.Utilities.UtilityFunctions;
 
-namespace Tests.PlayModeTests.MenuTests
+namespace Tests.PlayModeTests.PcMenuTests.LanguageMenuTests
 {
+    /// <summary>
+    /// Tests Main Menu language switch of all buttons and labels
+    /// </summary>
     [TestFixture]
-    public class PauseMenuLanguageTests : MenuLanguageTestBaseFixture
+    public class MainMenuLanguageTests : MenuLanguageTestBaseFixture
     {
-        public PauseMenuLanguageTests() : base(FallingCoilScenePath) {}
+        public MainMenuLanguageTests() : base(MainMenuScenePath) {}
 
-        [UnityTest, Order(1), Description("On selecting a new language in the Pause Menu, the LanguageManagers 'CurrentLanguage' property changes")]
+        [UnityTest, Order(1), Description("On selecting a new language in the Main Menu, the LanguageManagers 'CurrentLanguage' property changes")]
         public IEnumerator WhenSelectGermanLanguageThenLanguageManagerIsGerman()
         {
             yield return OpenLanguageSubMenu();
@@ -28,62 +31,61 @@ namespace Tests.PlayModeTests.MenuTests
                 $"LanguageManager's 'CurrentLanguage' property is not {ExpectedLanguage.ToString()}");
         }
 
-        private static TestCaseData[] _pauseMenuButtonKeysSource = new [] {
+        private static TestCaseData[] _mainMenuButtonKeysSource = new [] {
+            new TestCaseData("Menu Lab").Returns(null),
             new TestCaseData("Menu Audio").Returns(null),
             new TestCaseData("Menu Language").Returns(null),
-            new TestCaseData("Menu Network").Returns(null),
-            new TestCaseData("Menu Main Menu").Returns(null),
-            new TestCaseData("Menu Resume").Returns(null)
+            new TestCaseData("Menu Credits").Returns(null),
+            new TestCaseData("Menu Exit").Returns(null)
         };
         
-        [UnityTest, Order(2), TestCaseSource(nameof(_pauseMenuButtonKeysSource)),
-         Description("On selecting a new language in the Pause Menu, the Pause Menu's buttons match")]
-        public IEnumerator WhenSelectGermanLanguageThenPauseMenuButtonsAreGerman(string buttonKey)
+        [UnityTest, Order(2), TestCaseSource(nameof(_mainMenuButtonKeysSource)),
+         Description("On selecting a new language in the Main Menu, the Main Menu's buttons match")]
+        public IEnumerator WhenSelectGermanLanguageThenMainMenuButtonsAreGerman(string buttonKey)
         {
+            // Find and store top level main menu button
+            var button = GetButtonViaText(LanguageManager.Instance.GetString(buttonKey, DefaultLanguage));
+            
             yield return OpenLanguageSubMenu();
             yield return ClickGermanLanguageButton();
             
-            // Find top level pause menu button
-            var button = GetButtonViaText(LanguageManager.Instance.GetString(buttonKey, ExpectedLanguage));
-            
-            // Check if button text language matches
             var expectedButtonLabel = LanguageManager.Instance.GetString(buttonKey, ExpectedLanguage);
             var actualButtonLabel = button.GetComponentInChildren<TextMeshProUGUI>().text;
-            Assert.AreEqual(expectedButtonLabel, actualButtonLabel);
+            Assert.AreEqual(expectedButtonLabel, actualButtonLabel,
+                $"Button is wrong after changing language to {ExpectedLanguage}");
         }
 
-        [UnityTest, Order(3), Description("On selecting a new language in the Pause Menu, the main menu's title matches")]
-        public IEnumerator WhenSelectGermanLanguageThenPauseMenuTitleIsGerman()
+        [UnityTest, Order(3), Description("On selecting a new language in the Main Menu, the main menu's title matches")]
+        public IEnumerator WhenSelectGermanLanguageThenMainMenuTitleIsGerman()
         {
+            TextMeshProUGUI mainMenuTitle = GameObject.Find("preMenuColumnMainMenu(Clone)").GetComponentsInChildren<TextMeshProUGUI>()
+                .First(x => x.text == LanguageManager.Instance.GetString("Menu Main Menu", DefaultLanguage));
+            
             yield return OpenLanguageSubMenu();
             yield return ClickGermanLanguageButton();
             
-            // Find pause menu title
-            var labelKeyUnderTest = "Menu Pause Menu";
-            TextMeshProUGUI pauseMenuTitle = GameObject.Find("PauseMenu").GetComponentsInChildren<TextMeshProUGUI>()
-                .First(x => x.text == LanguageManager.Instance.GetString(labelKeyUnderTest, ExpectedLanguage));
-
-            // Check if pause menu title language matches
-            var expectedTitleLabel = LanguageManager.Instance.GetString(labelKeyUnderTest, ExpectedLanguage);
-            var actualTitleLabel = pauseMenuTitle.text;
+            var expectedTitleLabel = LanguageManager.Instance.GetString("Menu Main Menu", ExpectedLanguage);
+            var actualTitleLabel = mainMenuTitle.text;
             Assert.AreEqual(expectedTitleLabel, actualTitleLabel,
-                $"Pause Menu Title is wrong after changing language to {ExpectedLanguage}");
+                $"Main Menu Title is wrong after changing language to {ExpectedLanguage}");
         }
+
         
         private static TestCaseData[] _checkSubMenuTitlesSource = new [] {
+            new TestCaseData("Menu Lab", "preMenuColumnLaboratorySelection").Returns(null),
             new TestCaseData("Menu Audio", "preMenuColumnAudio").Returns(null),
             new TestCaseData("Menu Language", "preMenuColumnLanguage").Returns(null),
-            new TestCaseData("Menu Network", "preMenuColumnNetwork").Returns(null)
+            new TestCaseData("Menu Credits", "preMenuColumnCredits").Returns(null)
         };
         
         [UnityTest, Order(4), TestCaseSource(nameof(_checkSubMenuTitlesSource)),
-         Description("On selecting a new language in the Pause Menu, the sub menu's title matches")]
+         Description("On selecting a new language in the Main Menu, the sub menu's title matches")]
         public IEnumerator WhenSelectGermanLanguageThenSubMenuTitlesAreGerman(string titleKey, string menuPrefab)
         {
             yield return OpenLanguageSubMenu();
             yield return ClickGermanLanguageButton();
          
-            // Find and click sourced Pause Menu button
+            // Find and click sourced Main Menu button
             string sourceButtonLabel = LanguageManager.Instance.GetString(titleKey, ExpectedLanguage);
             GetButtonViaText(sourceButtonLabel).onClick.Invoke();
             yield return null;
@@ -95,17 +97,16 @@ namespace Tests.PlayModeTests.MenuTests
             var expectedTitleLabel = LanguageManager.Instance.GetString(titleKey, ExpectedLanguage);
             var actualTitleLabel = sourcedSubmenuTitle.text;
             
-            Assert.AreEqual(expectedTitleLabel, actualTitleLabel,
-                $"Title is wrong after changing language to {ExpectedLanguage}");
+            Assert.AreEqual(expectedTitleLabel, actualTitleLabel, $"Title is wrong after changing language to {ExpectedLanguage}");
         }
-        
-        [UnityTest, Order(5), Description("On selecting a new language in the Pause Menu, the Audio sub menu's labels match")]
+
+        [UnityTest, Order(5), Description("On selecting a new language in the Main Menu, the Audio sub menu's labels match")]
         public IEnumerator WhenSelectGermanLanguageThenAudioSubMenuIsGerman()
         {
             yield return OpenLanguageSubMenu();
             yield return ClickGermanLanguageButton();
             
-            // Find and click Pause Menu Audio button
+            // Find and click Main Menu Audio button
             string audioButtonLabel = LanguageManager.Instance.GetString("Menu Audio", ExpectedLanguage);
             GetButtonViaText(audioButtonLabel).onClick.Invoke();
             yield return null;
@@ -123,48 +124,29 @@ namespace Tests.PlayModeTests.MenuTests
                 $"FX Slider label is wrong after changing language to {ExpectedLanguage}");
         }
         
-        [UnityTest, Order(6), Description("On selecting a new language in the Pause Menu, the Language sub menu's labels match")]
+        [UnityTest, Order(6), Description("On selecting a new language in the Main Menu, the Language sub menu's labels match")]
         public IEnumerator WhenSelectGermanLanguageThenLanguageSubMenuIsGerman()
         {
             yield return OpenLanguageSubMenu();
             yield return ClickGermanLanguageButton();
-
-            // Find language selection buttons for English and German
+            
+            // Find English and German buttons
             string germanButtonLabel = LanguageManager.Instance.GetString("German", ExpectedLanguage);
             string englishButtonLabel = LanguageManager.Instance.GetString("English", ExpectedLanguage);
             Button germanButton = GetButtonViaText(germanButtonLabel);
             Button englishButton = GetButtonViaText(englishButtonLabel);
-
-            var expectedEnglishButtonLabel = LanguageManager.Instance.GetString("English", ExpectedLanguage);
-            var expectedGermanButtonLabel = LanguageManager.Instance.GetString("German", ExpectedLanguage);
             
-            // Get translated text labels from Buttons
+            var expectedGermanButtonLabel = LanguageManager.Instance.GetString("German", ExpectedLanguage);
             var actualGermanButtonLabel = germanButton.GetComponentInChildren<TextMeshProUGUI>().text;
+            
+            var expectedEnglishButtonLabel = LanguageManager.Instance.GetString("English", ExpectedLanguage);
             var actualEnglishButtonLabel = englishButton.GetComponentInChildren<TextMeshProUGUI>().text;
 
             Assert.AreEqual(expectedGermanButtonLabel, actualGermanButtonLabel,
                 $"German Language Button label is wrong after changing language to {ExpectedLanguage}");
+            
             Assert.AreEqual(expectedEnglishButtonLabel, actualEnglishButtonLabel,
                 $"English Language Button label is wrong after changing language to {ExpectedLanguage}");
-        }
-        
-        [UnityTest, Order(7), Description("On selecting a new language in the Pause Menu, the Network sub menu's labels match")]
-        public IEnumerator WhenSelectGermanLanguageThenNetworkSubMenuIsGerman()
-        {
-            yield return OpenLanguageSubMenu();
-            yield return ClickGermanLanguageButton();
-
-            // Find and click Pause Menu Audio button
-            string audioButtonLabel = LanguageManager.Instance.GetString("Menu Network", ExpectedLanguage);
-            GetButtonViaText(audioButtonLabel).onClick.Invoke();
-            yield return null;
-            
-            // Find buttons
-            
-            // Get translated text labels from Buttons
-            
-            // Assert
-            Assert.Fail("To be implemented");
         }
     }
 }
